Record rotation even if outputs are unrotated
Because this information is recorded in config dumps, it will mean that
rotation is explicitly set for each profile.

This behaviour is required for things to behave as expected when moving
from a profile with rotated outputs to one with those same outputs being
unrotated.

It also means that it can be used to revert changes made manually.